backend/requirements.txt
asgiref==3.4.1
Django==3.2.13
django-cors-headers==3.10.1
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.1
zope.event==4.5.0
zope.interface==5.4.0
